# docs/swagger.yaml
basePath: /v1
definitions:
  app_code.CodeType:
    enum:
    - success
    - user_not_found
    - password_error
    - service_error
    - auth_failed
    - wx_auth_failed
    - token_not_set
    - repeat
    - token_timeout
    - forbidden
    - user_exist
    - update_user_setting
    - get_user_setting
    - task_not_found
    - task_exist
    - bad_request
    - audit_param_invalid
    - not_found
    - task_run_not_found
    - create_task
    type: string
    x-enum-varnames:
    - Success
    - ErrorUserNotFound
    - ErrorUserPassword
    - ErrorServerError
    - ErrorAuthFailed
    - ErrorWxAuthFailed
    - ErrorTokenNotSet
    - ErrorRepeat
    - ErrorTokenTimeout
    - ErrorForbidden
    - ErrorUserExist
    - ErrorUpdateUserSetting
    - ErrorGetUserSetting
    - ErrorTaskNotFound
    - ErrorTaskExist
    - ErrorBadRequest
    - ErrorAuditParamInValid
    - ErrorNotFound
    - ErrorTaskRunNotFound
    - ErrorCreateTask
  entity.Notice:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  http_util.Response:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/app_code.CodeType'
        example: success
      data: {}
      error:
        example: error message
        type: string
    type: object
...
host: task.mytool.zone
info:
  contact: {}
  description: Task Mini Program API
  title: Task Mini Program
  version: "1.0"
paths:
  /notice/detail:
    get:
      consumes:
      - application/json
      description: 获取用户通知详情
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 通知ID
        in: query
        name: noticeID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Notice'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 获取用户通知详情
      tags:
      - 通知
  /notice/list:
    get:
      consumes:
      - application/json
      description: 获取用户通知列表
      parameters:
      - description: jwt_token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http_util.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Notice'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_util.Response'
      summary: 获取用户通知列表
      tags:
      - 通知
...
swagger: "2.0"